+++ src/tmp_test_file.ts
@@ -1,0 +1,11 @@
+export function testFuncA() {
+  console.log("Hello A");
+}
+
+export function testFuncB() {
+  console.log("Hello B");
+}
+
+export function testFuncC() {
+  console.log("Hello C");
+}